refine docs and check for ownership in products

+++ src/routes/products.ts
@@ -1,15 +1,14 @@
 import { zValidator } from '@hono/zod-validator'
 import { Hono } from 'hono'
 import { type JwtVariables, jwt } from 'hono/jwt'
-import { post_product } from '../validator/products'
+import { HTTPException } from 'hono/http-exception'
+import { post_product, put_product } from '../validator/products'
 
 import * as h from '../helpers/products'
 
 const products = new Hono<{ Variables: JwtVariables }>().use(
 	jwt({ secret: Bun.env.ACCESS_TOKEN_SECRET }),
 )
-
-//TODO: Decode jwt to get userid and check permissions.
 
 products.post('/', zValidator('json', post_product), async (c) => {
 	const id = Bun.randomUUIDv7()
@@ -40,9 +39,13 @@ products.get('/:id', async (c) => {
 	})
 })
 
-products.put('/:id', zValidator('json', post_product), async (c) => {
+products.put('/:id', zValidator('json', put_product), async (c) => {
 	const id = c.req.param('id')
 	const product_payload = c.req.valid('json')
+	const user_id = c.get('jwtPayload').sub
+
+	const { owner_id } = await h.get_product(id)
+	if (user_id !== owner_id) throw new HTTPException(403)
 
 	const result = await h.edit_product(id, product_payload)
 	return c.json({
@@ -53,6 +56,11 @@ products.put('/:id', zValidator('json', post_product), async (c) => {
 
 products.delete('/:id', async (c) => {
 	const id = c.req.param('id')
+	const user_id = c.get('jwtPayload').sub
+
+	const { owner_id } = await h.get_product(id)
+	if (user_id !== owner_id) throw new HTTPException(403)
+
 	const result = await h.delete_product(id)
 	return c.json({
 		message: 'Product deleted',
